Refactor: move RaftCore to raft_core.rs
